Add HW Two-dimensional array task 5

+++ Homework1/HWArrayClass.cs
@@ -190,7 +190,7 @@ namespace Homework4
             return newArray;
         }
 
-        public static void WriteArray(double[] array)
+        public static void WriteArray(int[] array)
         {
             for (int i = 0; i < array.Length; i++)
             {
